Install git-apps right after buildstep
This makes it so pushes will build the app just as soon as it's
safe to do so. If we installed git-apps *before* buildstep, a push
could trigger the buildstep buildhook before the image is finished
downloading.

This does mean any apps pushed *before* buildstep is installed
will not be built: see plushu/plushu-buildstep#4.